Change recording format and tool support.
Recordings now save frames with a timestamp rather than a delta. This
will make trimming recordings easier!

We now store a version number in the recording manifest. If it's not
there we assume a pre-v1 file and throw an error (unless compatibility
is forced - which makes no guarantees about actually working).