Add a date to `Message`s and use that for recordings.
There might be a delay or something in the recording process, but we
don't want frames to skip out of order - easy to just add a date to each
message.

Note that the value is stored as a JS date in serialised form, and as a
Unix timestamp in un/deserialised form.