Time-out recordings that have no recorded frames.
If a service broadcasts a manifest but no state frames, then its
recording file will sit in the "in progress" directory indefinitely.
This is sad. We therefore make a note of all data-less recordings on
each timeout pass, and if a recording still has no frames next time, we
consider it "finished".