In-progress recordings are stored unzipped.
When updating a manifest file stored in a zip, the only available method
is duplicating the zip file without the manifest, then adding the new
one. As recordings get longer, this process takes longer, meaning that
there's an ever-increasing CPU load from the DVR process.

By storing in-progress recordings on the filesystem, we eliminate that
problem, since we only create the ZIP file at the end. We also eliminate
the irritating "duplicate filename" warnings that `zipfile` was
generating.

There ought not to be a critical disk space impact from this change
(though of course there will be some).