Only clear DVR's cached analysis data when we're really finished.
If it's a session change, then in order to make sure the analysis data
isn't wiped by the time we try to write it, then we can just wait for
the running service to publish the updated data.

Still not convinced that there won't be a race condition there, but I'll
fix that if/when it starts being a pain.