Better handle when two services are scheduled too closely together.
If the first service is still running when a second instance of it needs
to be started, then we keep the first running and mark that it is
covering both scheduled events. The service will be kept running until
the end of the second event (assuming no other instances of that service
have been scheduled in that time period!)

If the second instance requires different arguments, then (since we
don't currently support multiple different instances of the same
service) we keep trying to start it until the first service has been
terminated.