Avoid concurrent calls to `writeState` in DVR recorder.
The only thing I can think of leading to repeated messages
appearing in iframes is that this method is being called for a second
time before `self.prevState` has been set; plausible as dumping the JSON
to disk is likely to take a significant amount of time.

I have no explanations as to why this only seems to happen for F1
recordings.