Support rawsocket endpoint URL, and disable SSL verification on Windows.
There are, apparently, ongoing problems with OpenSSL on Windows that
mean that for whatever reason the entirely valid certificate served by
the master router isn't trusted, and we can't connect.

Easiest, if not the most correct, course of action is to simply disable
SSL verification for this connection when running on Windows.