Publish analysis data more frequently.
Previously things weren't quite doing what they ought to; we were
looking at pending publishes once a minute, and only publishing things
that had previously been published more than `interval` ago at that
point.

Now, we check for pending publishes every second, and publish those that
meet the staleness criteria. We're still not going to be publishing
analysis data more frequently than `interval`, which is what was
intended; but now there won't be up to a minute of extra lag time beyond
that.

In addition, we use the minimum interval in standalone mode since we'll
be dealing with a very small number of direct connections, and therefore
don't need to worry so much about bandwidth for these updates.