Reuse stale UPnP port forwarding if one is present.
Unfortunately, it's a PITA to detect when a user terminates the plugin
by closing the terminal window directly (rather than Ctrl+C) and try to
handle that case. So we might end up leaving a stale port forwarding on
the user's router.

To somewhat mitigate that, we look for and reuse the port forwarding if
it still points to us.

An alternative might be to add an expiration to the port forward, but
apparently not all routers support that, so this is probably the best
approach.